pyproject.toml
[tool.poetry]
name = "git-railway"
version = "0.2.0"
description = "Visualise local git branches as neat interactive HTML pages"
authors = ["Gabriele N. Tornetta <phoenix1987@gmail.com>"]
license = "GPL-3.0-or-later"
readme = "README.md"
homepage = "https://github.com/P403n1x87/git-railway"
repository = "https://github.com/P403n1x87/git-railway"
keywords = ["tools", "scm"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
  "Programming Language :: Python :: 3.6",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
]
packages = [
  { include = "git_railway" },
]

[tool.poetry.scripts]
git-railway = "git_railway.__main__:main"

[tool.poetry.dependencies]
python = "^3.6"
GitPython = "^3.1.8"
svgwrite = "^1.4"
dataclasses = "*"
colour = "^0.1.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
